refactor: consolidate RPC envelope, error handling, and dispatch into shared layer
- Move RpcError to rpc.py and add make_error_response / check_response helpers
- Merge healthz, shutdown, and dispatch locking into RpcFacade.serve() closure
- Remove duplicated __init__, dispatch method, and _dispatch_lock from both
  HttpRpcServer and SocketRpcServer
- Remove dead req_id (sync RPC — no multiplexing) and no-op close() from clients
