+++ rpent/utils/__init__.py
@@ -1,9 +1,8 @@
 """Utility helpers: config, logging, path resolution, templates."""
 
 from rpent.utils.logging import get_logger, get_output_dir, init_output_dir
-from rpent.utils.rpc import RpcClient
+from rpent.utils.rpc import RpcClient, RpcError
 from rpent.utils.socket_rpc import (
-    RpcError,
     SocketRpcClient,
     SocketRpcServer,
 )
