+++ rpent/utils/rpc.py
@@ -13,6 +13,15 @@ if TYPE_CHECKING:
 logger = get_logger("rpc")
 
 
+class RpcError(RuntimeError):
+    """Raised when a remote method call returns an error."""
+
+    def __init__(self, method: str, message: str, *, traceback: str | None = None):
+        super().__init__(f"{method}: {message}")
+        self.method = method
+        self.server_traceback = traceback
+
+
 class RpcClient(Protocol):
     """Generic method-call RPC transport (agent → out-of-process server)."""
 
@@ -26,8 +35,24 @@ class RpcClient(Protocol):
     ) -> Any:
         """Invoke a remote method and return its result."""
 
-    def close(self) -> None:
-        """Release any client-side transport resources."""
+
+def make_error_response(exc: Exception) -> dict:
+    """Build the error envelope for a caught exception."""
+    import traceback as _tb
+    return {"ok": False, "error": str(exc), "traceback": _tb.format_exc()}
+
+
+def check_response(response: Any, method: str) -> Any:
+    """Validate RPC response envelope; raise ``RpcError`` on failure, return result."""
+    if not isinstance(response, dict):
+        raise RpcError(method, f"bad response type: {type(response).__name__}")
+    if not response.get("ok"):
+        raise RpcError(
+            method,
+            str(response.get("error", "<no error message>")),
+            traceback=response.get("traceback"),
+        )
+    return response.get("result")
 
 
 def wait_for_ready(
@@ -69,10 +94,8 @@ class RpcFacade:
     """Base class for subprocess RPC servers.
 
     Subclasses implement :meth:`_dispatch`; the base owns the shutdown
-    event, the ``shutdown`` RPC method, transport binding, parent-watch,
-    and clean teardown. ``healthz`` is answered by the transport itself
-    (see :class:`HttpRpcServer` / :class:`SocketRpcServer`), so subclasses
-    don't implement it either.
+    event, the ``shutdown`` / ``healthz`` RPC methods, transport binding,
+    parent-watch, and clean teardown.
 
     Usage::
 
@@ -91,8 +114,8 @@ class RpcFacade:
     def _dispatch(self, method: str, args: tuple, kwargs: dict) -> Any:
         """Business RPC dispatch. Override in subclasses.
 
-        Do not handle ``shutdown`` or ``healthz`` here — the base and the
-        transport framework take care of them.
+        Do not handle ``shutdown`` or ``healthz`` here — the base takes care
+        of them.
         """
         raise NotImplementedError
 
@@ -114,11 +137,16 @@ class RpcFacade:
         from rpent.utils.http_rpc import HttpRpcServer
         from rpent.utils.socket_rpc import SocketRpcServer
 
+        _lock = threading.Lock()
+
         def dispatch(method: str, args: tuple, kwargs: dict) -> Any:
+            if method == "healthz":
+                return {"status": "ok"}
             if method == "shutdown":
                 self._shutdown_event.set()
                 return {"ok": True}
-            return self._dispatch(method, args, kwargs)
+            with _lock:
+                return self._dispatch(method, args, kwargs)
 
         server_cls = HttpRpcServer if transport == "http" else SocketRpcServer
         server = server_cls((host, port), dispatch)
@@ -138,4 +166,11 @@ class RpcFacade:
             server.server_close()
 
 
-__all__ = ["RpcClient", "RpcFacade", "wait_for_ready"]
+__all__ = [
+    "RpcClient",
+    "RpcError",
+    "RpcFacade",
+    "check_response",
+    "make_error_response",
+    "wait_for_ready",
+]
